Adding using git for pulling dependencies

+++ src/janitor.rs
@@ -7,6 +7,7 @@ use std::fs::{
 use crate::config::Dependency;
 use crate::errors::MissingDependencies;
 use crate::lock::remove_lock;
+use crate::utils::get_download_tunnel;
 use crate::DEPENDENCY_DIR;
 
 // Health-check dependencies before we clean them, this one checks if they were unzipped
@@ -25,7 +26,8 @@ pub fn healthcheck_dependencies(dependencies: &[Dependency]) -> Result<(), Missi
 // Cleanup zips after the download
 pub fn cleanup_after(dependencies: &[Dependency]) -> Result<(), MissingDependencies> {
     for dependency in dependencies.iter() {
-        match cleanup_dependency(&dependency.name, &dependency.version, false) {
+        let via_git: bool = get_download_tunnel(&dependency.url) == "git";
+        match cleanup_dependency(&dependency.name, &dependency.version, false, via_git) {
             Ok(_) => {}
             Err(err) => {
                 println!("returning error {:?}", err);
@@ -57,18 +59,21 @@ pub fn cleanup_dependency(
     dependency_name: &str,
     dependency_version: &str,
     full: bool,
+    via_git: bool,
 ) -> Result<(), MissingDependencies> {
     let file_name: String = format!("{}-{}.zip", dependency_name, dependency_version);
     let new_path: std::path::PathBuf = DEPENDENCY_DIR.clone().join(file_name);
-    match remove_file(new_path) {
-        Ok(_) => {}
-        Err(_) => {
-            return Err(MissingDependencies::new(
-                dependency_name,
-                dependency_version,
-            ));
-        }
-    };
+    if !via_git {
+        match remove_file(new_path) {
+            Ok(_) => {}
+            Err(_) => {
+                return Err(MissingDependencies::new(
+                    dependency_name,
+                    dependency_version,
+                ));
+            }
+        };
+    }
     if full {
         let dir = DEPENDENCY_DIR.join(dependency_name);
         remove_dir_all(dir).unwrap();
@@ -118,7 +123,7 @@ mod tests {
             name: "@openzeppelin-contracts".to_string(),
             version: "2.3.0".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.3.0.zip".to_string(),
-        });
+            hash: String::new()});
         download_dependencies(&dependencies, false).await.unwrap();
         unzip_dependency(&dependencies[0].name, &dependencies[0].version).unwrap();
         healthcheck_dependency("@openzeppelin-contracts", "2.3.0").unwrap();
@@ -134,10 +139,10 @@ mod tests {
             name: "@openzeppelin-contracts".to_string(),
             version: "2.3.0".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.3.0.zip".to_string(),
-        });
+            hash: String::new() });
         download_dependencies(&dependencies, false).await.unwrap();
         unzip_dependency(&dependencies[0].name, &dependencies[0].version).unwrap();
-        cleanup_dependency("@openzeppelin-contracts", "2.3.0", false).unwrap();
+        cleanup_dependency("@openzeppelin-contracts", "2.3.0", false, false).unwrap();
     }
 
     #[test]
@@ -150,8 +155,14 @@ mod tests {
             name: "@openzeppelin-contracts".to_string(),
             version: "v-cleanup-nonexisting".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.3.0.zip".to_string(),
-        });
-        cleanup_dependency("@openzeppelin-contracts", "v-cleanup-nonexisting", false).unwrap_err();
+            hash: String::new()});
+        cleanup_dependency(
+            "@openzeppelin-contracts",
+            "v-cleanup-nonexisting",
+            false,
+            false,
+        )
+        .unwrap_err();
     }
 
     #[tokio::test]
@@ -164,12 +175,12 @@ mod tests {
             name: "@openzeppelin-contracts".to_string(),
             version: "2.3.0".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.3.0.zip".to_string(),
-        });
+            hash: String::new()});
         dependencies.push(Dependency {
             name: "@openzeppelin-contracts".to_string(),
             version: "2.4.0".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.4.0.zip".to_string(),
-        });
+            hash: String::new() });
 
         download_dependencies(&dependencies, false).await.unwrap();
         let _ = unzip_dependency(&dependencies[0].name, &dependencies[0].version);
@@ -188,7 +199,7 @@ mod tests {
             name: "@openzeppelin-contracts".to_string(),
             version: "cleanup-after-one-existing".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.3.0.zip".to_string(),
-        });
+            hash: String::new()});
 
         download_dependencies(&dependencies, false).await.unwrap();
         unzip_dependency(&dependencies[0].name, &dependencies[0].version).unwrap();
@@ -196,7 +207,7 @@ mod tests {
             name: "@openzeppelin-contracts".to_string(),
             version: "cleanup-after-one-existing-2".to_string(),
             url: "https://github.com/mario-eth/soldeer-versions/raw/main/all_versions/@openzeppelin-contracts~2.4.0.zip".to_string(),
-        });
+            hash: String::new()});
         match cleanup_after(&dependencies) {
             Ok(_) => {
                 assert_eq!("Invalid State", "");
